revert(terminal): restore stable IME input path
Reverts the terminal input changes from #386, #389, and #391 that regressed Korean IME Space handling in v1.13.4+.

Keeps unrelated fixes and release metadata intact.
